issue-29 Create Clikt program for closing prs
