+++ src/main/kotlin/pullrequestfactory/io/clikt/CliktClosePrsPgram.kt
@@ -1,0 +1,32 @@
+package pullrequestfactory.io.clikt
+
+import pullrequestfactory.domain.Candidate
+import pullrequestfactory.domain.GithubPRFactory
+import pullrequestfactory.domain.branches.BranchSyntaxValidator
+import pullrequestfactory.domain.pullrequests.PullRequestLastNotFinishedMarker
+import pullrequestfactory.io.programs.Program
+import pullrequestfactory.io.repositories.GithubHttpBranchesRepos
+import pullrequestfactory.io.repositories.GithubHttpPullRequestsRepo
+import pullrequestfactory.io.repositories.KhttpClientStats
+import pullrequestfactory.io.uis.ConsoleUI
+
+class CliktClosePrsPgram(
+        private val ui: ConsoleUI,
+        private val repoUrl: String,
+        private val httpClient: KhttpClientStats,
+        private val candidate: Candidate
+) : Program {
+
+    override fun execute() {
+        val branchesRepo = GithubHttpBranchesRepos(repoUrl, ui, httpClient)
+        val prRepo = GithubHttpPullRequestsRepo(repoUrl, ui, httpClient)
+        val f = GithubPRFactory(
+                ui,
+                branchesRepo,
+                prRepo,
+                BranchSyntaxValidator(ui),
+                PullRequestLastNotFinishedMarker())
+        f.close_pull_requests_for(candidate)
+    }
+
+}
